Merge dev (#21)
* Adding error management to lexer, and removes some checks from lexer

* Lexer-Parser v2

This version of the lexer parser proceed to contextual analysis only at
the parsing step. In a following commit, an error system will be added.

* New error management system

Adding a new error management system in the parser step. Every error
spotted at parsing step will stop the AST generation.

* Parser is done

Roughly the parser is complete

* fixes leaks in assembly parsing

* Adding whitespaces to the tokenizer

Now the tokenizer sees whitespaces (only space and tab) as tokens.
It will make the AST aware of the exact surroundings of each token
and therefore a better understanding of how to interpret the
input.

* Parser v3

Adding better whitespace management to the AST building and fixes
all leaks in the parser. The tokenizer is also improved and
refactored.

* Add color support for macOS

* Add support of macOS for the program name variable of GLIBC